Create new reference when same name and different text
When two references have the same name but different text, initial code
displays an error. Here we choose a different behavior, by appending '-1'
(or '-2' etc. if needed) to the name and so creating a new reference. The
error is displayed only when a reference with empty content has the same
name (reuse of existing reference) as these duplicates.
French and English translations or error message are changed.